CartPoleHighRewardIn2026/requirements.txt
gymnasium==0.29.1
torch==2.1.0
numpy==1.24.3
matplotlib==3.7.1
tensorboard==2.14.0
pyyaml==6.0
tqdm==4.65.0